Move Torch hook into its own module
