+++ syft/frameworks/torch/__init__.py
@@ -1,6 +1,6 @@
-from . import hook_args
 from . import tensors
 from . import federated
 from . import differential_privacy
 from . import crypto
+from .hook import hook_args
 from .hook import TorchHook
