+++ syft/frameworks/torch/hook/__init__.py
@@ -1,0 +1,1 @@
+from .hook import TorchHook
