+++ test/torch/test_hook.py
@@ -47,7 +47,7 @@ def test_pointer_found_exception(workers):
 
 
 def test_build_get_child_type():
-    from syft.frameworks.torch.hook_args import build_rule, build_get_tensor_type
+    from syft.frameworks.torch.hook.hook_args import build_rule, build_get_tensor_type
 
     x = torch.Tensor([1, 2, 3])
     args = (x, [[1, x]])
